+++ src/lsp4clj/server.clj
@@ -247,7 +247,11 @@
      :pending-requests* (atom {})
      :join (promise)}))
 
+#_{:clj-kondo/ignore [:clojure-lsp/unused-public-var]}
 (defn stdio-server
+  "DEPRECATED: Will be removed in a future release. Prefer
+  lsp4clj.io-server/server or lsp4clj.io-server/stdio-server."
+  {:deprecated "Use lsp4clj.io-server/server instead"}
   [{:keys [in out] :as opts}]
   (chan-server (assoc opts
                       :input-ch (json-rpc/input-stream->input-chan in)