+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support communication through a socket, as an alternative to stdio. #1
+- Deprecate `lsp4clj.server/stdio-server`. Use the identical `lsp4clj.io-server/server` or `lsp4clj.io-server/stdio-server` instead.
 
 ## v1.0.1